Port diagnostic::on_const
