+++ compiler/rustc_attr_parsing/src/attributes/diagnostic/mod.rs
@@ -1,30 +1,286 @@
-#![allow(warnings)]
 use std::ops::Range;
 
 use rustc_errors::E0232;
 use rustc_hir::AttrPath;
 use rustc_hir::attrs::diagnostic::{
-    FilterFormatString, Flag, FormatArg, FormatString, LitOrArg, Name, NameValue,
-    OnUnimplementedCondition, Piece, Predicate,
+    AppendConstMessage, Directive, FilterFormatString, Flag, FormatArg, FormatString, LitOrArg,
+    Name, NameValue, OnUnimplementedCondition, Piece, Predicate,
 };
-use rustc_hir::lints::FormatWarning;
+use rustc_hir::lints::{AttributeLintKind, FormatWarning};
 use rustc_macros::Diagnostic;
 use rustc_parse_format::{
     Argument, FormatSpec, ParseError, ParseMode, Parser, Piece as RpfPiece, Position,
 };
+use rustc_session::lint::builtin::{
+    MALFORMED_DIAGNOSTIC_ATTRIBUTES, MALFORMED_DIAGNOSTIC_FORMAT_LITERALS,
+};
 use rustc_span::{Ident, InnerSpan, Span, Symbol, kw, sym};
-use thin_vec::ThinVec;
+use thin_vec::{ThinVec, thin_vec};
 
+use crate::context::{AcceptContext, Stage};
 use crate::parser::{ArgParser, MetaItemListParser, MetaItemOrLitParser, MetaItemParser};
 
-pub mod on_unimplemented;
+pub(crate) mod on_const;
+pub(crate) mod on_unimplemented;
 
 #[derive(Copy, Clone)]
 pub(crate) enum Mode {
-    // `#[rustc_on_unimplemented]`
+    /// `#[rustc_on_unimplemented]`
     RustcOnUnimplemented,
-    // `#[diagnostic::...]`
+    /// `#[diagnostic::on_unimplemented]`
     DiagnosticOnUnimplemented,
+    /// `#[diagnostic::on_const]`
+    DiagnosticOnConst,
+}
+
+fn merge_directives<S: Stage>(
+    cx: &mut AcceptContext<'_, '_, S>,
+    first: &mut Option<(Span, Directive)>,
+    later: (Span, Directive),
+) {
+    if let Some((_, first)) = first {
+        if first.is_rustc_attr || later.1.is_rustc_attr {
+            cx.emit_err(DupesNotAllowed);
+        }
+
+        merge(cx, &mut first.message, later.1.message, sym::message);
+        merge(cx, &mut first.label, later.1.label, sym::label);
+        first.notes.extend(later.1.notes);
+    } else {
+        *first = Some(later);
+    }
+}
+
+fn merge<T, S: Stage>(
+    cx: &mut AcceptContext<'_, '_, S>,
+    first: &mut Option<(Span, T)>,
+    later: Option<(Span, T)>,
+    option_name: Symbol,
+) {
+    match (first, later) {
+        (Some(_) | None, None) => {}
+        (Some((first_span, _)), Some((later_span, _))) => {
+            cx.emit_lint(
+                MALFORMED_DIAGNOSTIC_ATTRIBUTES,
+                AttributeLintKind::IgnoredDiagnosticOption {
+                    first_span: *first_span,
+                    later_span,
+                    option_name,
+                },
+                later_span,
+            );
+        }
+        (first @ None, Some(later)) => {
+            first.get_or_insert(later);
+        }
+    }
+}
+
+fn parse_directive_items<'p, S: Stage>(
+    cx: &mut AcceptContext<'_, '_, S>,
+    mode: Mode,
+    items: impl Iterator<Item = &'p MetaItemOrLitParser>,
+    is_root: bool,
+) -> Option<Directive> {
+    let condition = None;
+    let mut message: Option<(Span, _)> = None;
+    let mut label: Option<(Span, _)> = None;
+    let mut notes = ThinVec::new();
+    let mut parent_label = None;
+    let mut subcommands = ThinVec::new();
+    let mut append_const_msg = None;
+
+    for item in items {
+        let span = item.span();
+
+        macro malformed() {{
+            match mode {
+                Mode::RustcOnUnimplemented => {
+                    cx.emit_err(NoValueInOnUnimplemented { span: item.span() });
+                }
+                Mode::DiagnosticOnUnimplemented => {
+                    cx.emit_lint(
+                        MALFORMED_DIAGNOSTIC_ATTRIBUTES,
+                        AttributeLintKind::MalformedOnUnimplementedAttr { span },
+                        span,
+                    );
+                }
+                Mode::DiagnosticOnConst => {
+                    cx.emit_lint(
+                        MALFORMED_DIAGNOSTIC_ATTRIBUTES,
+                        AttributeLintKind::MalformedOnConstAttr { span },
+                        span,
+                    );
+                }
+            }
+            continue;
+        }}
+
+        macro or_malformed($($code:tt)*) {{
+            let Some(ret) = (try {
+                $($code)*
+            }) else {
+
+                malformed!()
+            };
+            ret
+        }}
+
+        macro duplicate($name: ident, $($first_span:tt)*) {{
+            match mode {
+                Mode::RustcOnUnimplemented => {
+                    cx.emit_err(NoValueInOnUnimplemented { span: item.span() });
+                }
+                Mode::DiagnosticOnUnimplemented |Mode::DiagnosticOnConst => {
+                    cx.emit_lint(
+                        MALFORMED_DIAGNOSTIC_ATTRIBUTES,
+                        AttributeLintKind::IgnoredDiagnosticOption {
+                            first_span: $($first_span)*,
+                            later_span: span,
+                            option_name: $name,
+                        },
+                        span,
+                    );
+                }
+            }
+        }}
+
+        let item: &MetaItemParser = or_malformed!(item.meta_item()?);
+        let name = or_malformed!(item.ident()?).name;
+
+        // Some things like `message = "message"` must have a value.
+        // But with things like `append_const_msg` that is optional.
+        let value: Option<Ident> = match item.args().name_value() {
+            Some(nv) => Some(or_malformed!(nv.value_as_ident()?)),
+            None => None,
+        };
+
+        let mut parse_format = |input: Ident| {
+            let snippet = cx.sess.source_map().span_to_snippet(input.span).ok();
+            let is_snippet = snippet.is_some();
+            match parse_format_string(input.name, snippet, input.span, mode) {
+                Ok((f, warnings)) => {
+                    for warning in warnings {
+                        let (FormatWarning::InvalidSpecifier { span, .. }
+                        | FormatWarning::PositionalArgument { span, .. }) = warning;
+                        cx.emit_lint(
+                            MALFORMED_DIAGNOSTIC_FORMAT_LITERALS,
+                            AttributeLintKind::MalformedDiagnosticFormat { warning },
+                            span,
+                        );
+                    }
+
+                    f
+                }
+                Err(e) => {
+                    cx.emit_lint(
+                        MALFORMED_DIAGNOSTIC_FORMAT_LITERALS,
+                        AttributeLintKind::DiagnosticWrappedParserError {
+                            description: e.description,
+                            label: e.label,
+                            span: slice_span(input.span, e.span, is_snippet),
+                        },
+                        input.span,
+                    );
+                    // We could not parse the input, just use it as-is.
+                    FormatString {
+                        input: input.name,
+                        span: input.span,
+                        pieces: thin_vec![Piece::Lit(input.name)],
+                    }
+                }
+            }
+        };
+        match (mode, name) {
+            (_, sym::message) => {
+                let value = or_malformed!(value?);
+                if let Some(message) = &message {
+                    duplicate!(name, message.0)
+                } else {
+                    message = Some((item.span(), parse_format(value)));
+                }
+            }
+            (_, sym::label) => {
+                let value = or_malformed!(value?);
+                if let Some(label) = &label {
+                    duplicate!(name, label.0)
+                } else {
+                    label = Some((item.span(), parse_format(value)));
+                }
+            }
+            (_, sym::note) => {
+                let value = or_malformed!(value?);
+                notes.push(parse_format(value))
+            }
+
+            (Mode::RustcOnUnimplemented, sym::append_const_msg) => {
+                append_const_msg = if let Some(msg) = value {
+                    Some(AppendConstMessage::Custom(msg.name, item.span()))
+                } else {
+                    Some(AppendConstMessage::Default)
+                }
+            }
+            (Mode::RustcOnUnimplemented, sym::parent_label) => {
+                let value = or_malformed!(value?);
+                if parent_label.is_none() {
+                    parent_label = Some(parse_format(value));
+                } else {
+                    // warn
+                }
+            }
+            (Mode::RustcOnUnimplemented, sym::on) => {
+                if is_root {
+                    let items = or_malformed!(item.args().list()?);
+                    let mut iter = items.mixed();
+                    let condition: &MetaItemOrLitParser = match iter.next() {
+                        Some(c) => c,
+                        None => {
+                            cx.emit_err(InvalidOnClause::Empty { span });
+                            continue;
+                        }
+                    };
+
+                    let condition = parse_condition(condition);
+
+                    if items.len() < 2 {
+                        // Something like `#[rustc_on_unimplemented(on(.., /* nothing */))]`
+                        // There's a condition but no directive behind it, this is a mistake.
+                        malformed!();
+                    }
+
+                    let mut directive =
+                        or_malformed!(parse_directive_items(cx, mode, iter, false)?);
+
+                    match condition {
+                        Ok(c) => {
+                            directive.condition = Some(c);
+                            subcommands.push(directive);
+                        }
+                        Err(e) => {
+                            cx.emit_err(e);
+                        }
+                    }
+                } else {
+                    malformed!();
+                }
+            }
+
+            _other => {
+                malformed!();
+            }
+        }
+    }
+
+    Some(Directive {
+        is_rustc_attr: matches!(mode, Mode::RustcOnUnimplemented),
+        condition,
+        subcommands,
+        message,
+        label,
+        notes,
+        parent_label,
+        append_const_msg,
+    })
 }
 
 pub(crate) fn parse_format_string(
@@ -74,14 +330,8 @@ fn parse_arg(
             (Mode::RustcOnUnimplemented { .. }, sym::This) => FormatArg::This,
             (Mode::RustcOnUnimplemented { .. }, sym::Trait) => FormatArg::Trait,
             // Any attribute can use these
-            (
-                Mode::RustcOnUnimplemented { .. } | Mode::DiagnosticOnUnimplemented { .. },
-                kw::SelfUpper,
-            ) => FormatArg::SelfUpper,
-            (
-                Mode::RustcOnUnimplemented { .. } | Mode::DiagnosticOnUnimplemented { .. },
-                generic_param,
-            ) => FormatArg::GenericParam { generic_param, span },
+            (_, kw::SelfUpper) => FormatArg::SelfUpper,
+            (_, generic_param) => FormatArg::GenericParam { generic_param, span },
         },
 
         // `{:1}` and `{}` are ignored
@@ -163,7 +413,7 @@ fn parse_predicate(input: &MetaItemOrLitParser) -> Result<Predicate, InvalidOnCl
             let Some(value) = p.value_as_ident() else {
                 return Err(InvalidOnClause::UnsupportedLiteral { span: p.args_span() });
             };
-            let name = parse_name(predicate);
+            let name = parse_name(predicate.name);
             let value = parse_filter(value.name);
             let kv = NameValue { name, value };
             Ok(Predicate::Match(kv))
@@ -190,7 +440,7 @@ fn parse_flag(Ident { name, span }: Ident) -> Result<Flag, InvalidOnClause> {
     }
 }
 
-fn parse_name(Ident { name, span }: Ident) -> Name {
+fn parse_name(name: Symbol) -> Name {
     match name {
         kw::SelfUpper => Name::SelfUpper,
         sym::from_desugaring => Name::FromDesugaring,
@@ -211,14 +461,11 @@ fn parse_filter(input: Symbol) -> FilterFormatString {
                 // from the `Display` impl of `InferTy` to be precise.
                 //
                 // Don't try to format these later!
-                Position::ArgumentNamed(arg @ "integer" | arg @ "integral" | arg @ "float") => {
+                Position::ArgumentNamed(arg @ ("integer" | "integral" | "float")) => {
                     LitOrArg::Lit(Symbol::intern(&format!("{{{arg}}}")))
                 }
 
-                // FIXME(mejrs) We should check if these correspond to a generic of the trait.
                 Position::ArgumentNamed(arg) => LitOrArg::Arg(Symbol::intern(arg)),
-
-                // FIXME(mejrs) These should really be warnings/errors
                 Position::ArgumentImplicitlyIs(_) => LitOrArg::Lit(sym::empty_braces),
                 Position::ArgumentIs(idx) => LitOrArg::Lit(Symbol::intern(&format!("{{{idx}}}"))),
             },
@@ -270,21 +517,12 @@ pub(crate) enum InvalidOnClause {
         span: Span,
         invalid_flag: Symbol,
     },
-    #[diag("invalid name in `on`-clause", code = E0232)]
-    InvalidName {
-        #[primary_span]
-        #[label(
-            "expected one of `cause`, `from_desugaring`, `Self` or any generic parameter of the trait, not `{$invalid_name}`"
-        )]
-        span: Span,
-        invalid_name: Symbol,
-    },
 }
 
 #[derive(Diagnostic)]
 #[diag("this attribute must have a value", code = E0232)]
 #[note("e.g. `#[rustc_on_unimplemented(message=\"foo\")]`")]
-pub struct NoValueInOnUnimplemented {
+pub(crate) struct NoValueInOnUnimplemented {
     #[primary_span]
     #[label("expected value here")]
     pub span: Span,
@@ -294,4 +532,4 @@ pub struct NoValueInOnUnimplemented {
 #[diag(
     "using multiple `rustc_on_unimplemented` (or mixing it with `diagnostic::on_unimplemented`) is not supported"
 )]
-pub struct DupesNotAllowed;
+pub(crate) struct DupesNotAllowed;
