+++ tests/ui/unpretty/diagnostic-attr.stdout
@@ -1,11 +1,12 @@
-extern crate std;
-#[attr = PreludeImport]
-use ::std::prelude::rust_2015::*;
 //@ compile-flags: -Zunpretty=hir
 //@ check-pass
 //@ edition: 2015
+#![feature(diagnostic_on_const)]
+extern crate std;
+#[attr = PreludeImport]
+use ::std::prelude::rust_2015::*;
 
-#[attr = OnUnimplemented {directive: OnUnimplementedDirective {is_rustc_attr: false,
+#[attr = OnUnimplemented {directive: Directive {is_rustc_attr: false,
 subcommands: [],
 message: FormatString {input: "My Message for `ImportantTrait<{A}>` implemented for `{Self}`",
 pieces: [Lit("My Message for `ImportantTrait<"),
@@ -18,3 +19,14 @@ trait ImportantTrait<A> { }
 
 #[attr = DoNotRecommend]
 impl <T> ImportantTrait<T> for T where T: Clone { }
+
+struct X;
+
+#[attr = OnConst {directive: Directive {is_rustc_attr: false, subcommands: [],
+message: FormatString {input: "My Message for `ImportantTrait<u8>` implemented for `{Self}`",
+pieces: [Lit("My Message for `ImportantTrait<u8>` implemented for `"),
+Arg(SelfUpper), Lit("`")]}, label: FormatString {input: "My Label",
+pieces: [Lit("My Label")]}, notes: [FormatString {input: "Note 1",
+pieces: [Lit("Note 1")]}, FormatString {input: "Note 2",
+pieces: [Lit("Note 2")]}]}}]
+impl ImportantTrait<u8> for X { }
